+++ android/src/main/kotlin/io/intheloup/geolocation/helper/Logger.kt
@@ -1,4 +1,4 @@
-package io.alfanhui.new_geolocation.helper
+package io.intheloup.geolocation.helper
 
 private const val tag = "geolocation"